cancer-detection-ai/requirements.txt
google-cloud-aiplatform
google-cloud-storage
pandas
numpy
matplotlib
seaborn
tensorflow
monai
scikit-learn
shap